+++ docs/CHANGELOG.md
@@ -1,8 +1,7 @@
 ## V 1.0 - in progress:
 - Fully functional today's tasks ✔
-- Month view allowing you to see individual days with their assigned tasks
+- Month view allowing you to see individual days with their assigned tasks ✔
 - Functional settings
-- Sharing notes with other people
 
 ## V 0.1.0:
 - Working login ✔